refactor: "수강신청한다" 책임 이동에 따른 테스트 수정

+++ src/main/java/nextstep/users/domain/NsUser.java
@@ -87,7 +87,7 @@ public class NsUser {
         return payment;
     }
 
-    public Payment findPaymentBySessionId(int sessionId) {
+    public Payment findPaymentBySessionId(long sessionId) {
         return this.payments.stream()
                 .filter(payment -> payment.isSameSessionId(sessionId))
                 .findAny()
